backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.18
django-cors-headers==4.4.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djoser==2.2.3
flake8==7.1.1
idna==3.8
mccabe==0.7.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
